pyproject.toml
[project]
name = "stylematch"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "streamlit>=1.35",
    "pillow>=10",
    "python-dotenv",
    "pydantic>=2",
]

[project.optional-dependencies]
dev = ["pytest"]
